refactor(webrtc): split webrtc_manager.cpp into webrtc_protocol + manager
- Create webrtc_protocol.h/cpp with all 23 protocol/helper functions
  from the anonymous namespace (protocol conversion, response population,
  session utilities, filter parsing, transport, SDP/codec utilities)
- Add BuildManualCandidateSdp and WaitForSdpAnswer implementations
- Refactor webrtc_manager.cpp: remove anonymous namespace, extract
  SetupPeerConnectionCallbacks, SetupMediaTracks, SetupDataChannels,
  HandleControlMessage, HandleProcessingMessage, HandleVideoFrame as
  private methods making CreateSession an ~100-line orchestrator
- Fix initialization order bug: create memory_pool before LiveVideoProcessor
- Add webrtc_protocol cc_library to BUILD with proper deps
- Add #include <future> to webrtc_manager.h
- Fix dangling-reference warnings in WaitForSdpAnswer (copy optional value)
